+++ src/fragments/pages.ts
@@ -1,5 +1,7 @@
 import gql from "graphql-tag";
 
+import { metadataFragment } from "./metadata";
+
 export const pageFragment = gql`
   fragment PageFragment on Page {
     id
@@ -11,8 +13,10 @@ export const pageFragment = gql`
 
 export const pageDetailsFragment = gql`
   ${pageFragment}
+  ${metadataFragment}
   fragment PageDetailsFragment on Page {
     ...PageFragment
+    ...MetadataFragment
     contentJson
     seoTitle
     seoDescription