+++ src/orders/views/OrderDetails/index.tsx
@@ -23,25 +23,16 @@ import React from "react";
 import { useIntl } from "react-intl";
 
 import { customerUrl } from "../../../customers/urls";
-import {
-  getMutationState,
-  getStringOrPlaceholder,
-  maybe,
-  transformAddressToForm
-} from "../../../misc";
+import { getMutationState, getStringOrPlaceholder, maybe } from "../../../misc";
 import { productUrl } from "../../../products/urls";
 import {
   FulfillmentStatus,
   JobStatusEnum,
   OrderStatus
 } from "../../../types/globalTypes";
-import OrderAddressEditDialog from "../../components/OrderAddressEditDialog";
 import OrderCancelDialog from "../../components/OrderCancelDialog";
 import OrderDetailsPage from "../../components/OrderDetailsPage";
 import OrderDraftCancelDialog from "../../components/OrderDraftCancelDialog/OrderDraftCancelDialog";
-import OrderDraftFinalizeDialog, {
-  OrderDraftFinalizeWarning
-} from "../../components/OrderDraftFinalizeDialog";
 import OrderDraftPage from "../../components/OrderDraftPage";
 import OrderFulfillmentCancelDialog from "../../components/OrderFulfillmentCancelDialog";
 import OrderFulfillmentTrackingDialog from "../../components/OrderFulfillmentTrackingDialog";
@@ -52,7 +43,6 @@ import OrderProductAddDialog from "../../components/OrderProductAddDialog";
 import OrderShippingMethodEditDialog from "../../components/OrderShippingMethodEditDialog";
 import OrderOperations from "../../containers/OrderOperations";
 import { TypedOrderDetailsQuery, useOrderVariantSearch } from "../../queries";
-import { OrderDetails_order } from "../../types/OrderDetails";
 import {
   orderDraftListUrl,
   orderFulfillUrl,
@@ -61,37 +51,8 @@ import {
   OrderUrlDialog,
   OrderUrlQueryParams
 } from "../../urls";
+import OrderAddressFields from "./OrderAddressFields";
 import { OrderDetailsMessages } from "./OrderDetailsMessages";
-
-const orderDraftFinalizeWarnings = (order: OrderDetails_order) => {
-  const warnings = [] as OrderDraftFinalizeWarning[];
-  if (!(order && order.shippingAddress)) {
-    warnings.push(OrderDraftFinalizeWarning.NO_SHIPPING);
-  }
-  if (!(order && order.billingAddress)) {
-    warnings.push(OrderDraftFinalizeWarning.NO_BILLING);
-  }
-  if (!(order && (order.user || order.userEmail))) {
-    warnings.push(OrderDraftFinalizeWarning.NO_USER);
-  }
-  if (
-    order &&
-    order.lines &&
-    order.lines.filter(line => line.isShippingRequired).length > 0 &&
-    order.shippingMethod === null
-  ) {
-    warnings.push(OrderDraftFinalizeWarning.NO_SHIPPING_METHOD);
-  }
-  if (
-    order &&
-    order.lines &&
-    order.lines.filter(line => line.isShippingRequired).length === 0 &&
-    order.shippingMethod !== null
-  ) {
-    warnings.push(OrderDraftFinalizeWarning.UNNECESSARY_SHIPPING_METHOD);
-  }
-  return warnings;
-};
 
 interface OrderDetailsProps {
   id: string;
@@ -513,7 +474,9 @@ export const OrderDetails: React.FC<OrderDetailsProps> = ({ id, params }) => {
                               input: data
                             })
                           }
-                          onDraftFinalize={() => openModal("finalize")}
+                          onDraftFinalize={() =>
+                            orderDraftFinalize.mutate({ id })
+                          }
                           onDraftRemove={() => openModal("cancel")}
                           onOrderLineAdd={() => openModal("add-order-line")}
                           onBack={() => navigate(orderDraftListUrl())}
@@ -560,18 +523,6 @@ export const OrderDetails: React.FC<OrderDetailsProps> = ({ id, params }) => {
                           onConfirm={() => orderDraftCancel.mutate({ id })}
                           open={params.action === "cancel"}
                           orderNumber={getStringOrPlaceholder(order?.number)}
-                        />
-                        <OrderDraftFinalizeDialog
-                          confirmButtonState={orderDraftFinalize.opts.status}
-                          errors={
-                            orderDraftFinalize.opts.data?.draftOrderComplete
-                              .errors || []
-                          }
-                          onClose={closeModal}
-                          onConfirm={() => orderDraftFinalize.mutate({ id })}
-                          open={params.action === "finalize"}
-                          orderNumber={getStringOrPlaceholder(order?.number)}
-                          warnings={orderDraftFinalizeWarnings(order)}
                         />
                         <OrderShippingMethodEditDialog
                           confirmButtonState={
@@ -623,49 +574,14 @@ export const OrderDetails: React.FC<OrderDetailsProps> = ({ id, params }) => {
                         />
                       </>
                     )}
-                    <OrderAddressEditDialog
-                      confirmButtonState={orderUpdate.opts.status}
-                      address={transformAddressToForm(order?.shippingAddress)}
-                      countries={
-                        data?.shop?.countries.map(country => ({
-                          code: country.code,
-                          label: country.country
-                        })) || []
-                      }
-                      errors={orderUpdate.opts.data?.orderUpdate.errors || []}
-                      open={params.action === "edit-shipping-address"}
-                      variant="shipping"
+                    <OrderAddressFields
+                      isDraft={order?.status === OrderStatus.DRAFT}
+                      orderUpdate={orderUpdate}
+                      orderDraftUpdate={orderDraftUpdate}
+                      data={data}
+                      id={id}
                       onClose={closeModal}
-                      onConfirm={shippingAddress =>
-                        orderUpdate.mutate({
-                          id,
-                          input: {
-                            shippingAddress
-                          }
-                        })
-                      }
-                    />
-                    <OrderAddressEditDialog
-                      confirmButtonState={orderUpdate.opts.status}
-                      address={transformAddressToForm(order?.billingAddress)}
-                      countries={
-                        data?.shop?.countries.map(country => ({
-                          code: country.code,
-                          label: country.country
-                        })) || []
-                      }
-                      errors={orderUpdate.opts.data?.orderUpdate.errors || []}
-                      open={params.action === "edit-billing-address"}
-                      variant="billing"
-                      onClose={closeModal}
-                      onConfirm={billingAddress =>
-                        orderUpdate.mutate({
-                          id,
-                          input: {
-                            billingAddress
-                          }
-                        })
-                      }
+                      action={params.action}
                     />
                   </>
                 )}