+++ src/pages/types/PageDetails.ts
@@ -6,12 +6,26 @@
 // GraphQL query operation: PageDetails
 // ====================================================
 
+export interface PageDetails_page_metadata {
+  __typename: "MetadataItem";
+  key: string;
+  value: string;
+}
+
+export interface PageDetails_page_privateMetadata {
+  __typename: "MetadataItem";
+  key: string;
+  value: string;
+}
+
 export interface PageDetails_page {
   __typename: "Page";
   id: string;
   title: string;
   slug: string;
   isPublished: boolean;
+  metadata: (PageDetails_page_metadata | null)[];
+  privateMetadata: (PageDetails_page_privateMetadata | null)[];
   contentJson: any;
   seoTitle: string | null;
   seoDescription: string | null;